Refactor edit on click logic into a hook
